+++ src/features/calendar/components/MonthViewCalendar/hooks/useTimeSelection.js
@@ -1,0 +1,10 @@
+import { useCalendarStore } from "../../../../../store/index.ts";
+
+export const useTimeSelection = () => {
+  const { startTime, endTime } = useCalendarStore();
+
+  return {
+    startTime,
+    endTime,
+  };
+};
